Semi-Final Update
Update

+++ SelectGroup.php
@@ -7,14 +7,14 @@ if(array_key_exists('advisorID', $_SESSION)){
 ?>
 <h2>Select Group</h2>
 <?php
-    if(!array_key_exists('date', $_POST)){  
+if(!array_key_exists('date', $_POST)){  
 ?>
 <form id="weekForm" action="SelectGroup.php" method="post">
 <?php include 'includes/selectWeek.php'; ?>
 </form>
 <?php
     }
-    if(array_key_exists('week', $_POST)){
+    if(array_key_exists('week', $_POST) && !array_key_exists('date', $_POST)){
         echo "<form action=\"SelectGroup.php\" method=\"post\">";
         echo "<br><div id=\"selectTitle\">Select day:</div>";
         $week = $CALENDAR->weeks[(int)$_POST['week']];
@@ -24,9 +24,14 @@ if(array_key_exists('advisorID', $_SESSION)){
           echo "<li><input type=\"radio\" name=\"date\" value=\"".$date."\">";
           echo date_to_string($date)."</li>";
         }
-        echo "</ul></div>";  
+        echo "</ul></div>";       
         echo "<div id=\"submit\"><input type=\"submit\" name=\"submitDay\">";
         echo "</div>";
+        echo "<input type=\"hidden\" name=\"week\" value=\"".$_POST['week']."\">";
+        if(array_key_exists('submitDay', $_POST)){
+            echo "<div id=\"error\"><img src=\"includes/error.png\" id=\"errorImg\">";
+            echo "Please select a date</div>";   
+        }
     }
     
     else if (array_key_exists('date', $_POST) && !array_key_exists('time', $_POST) ){
@@ -35,29 +40,43 @@ if(array_key_exists('advisorID', $_SESSION)){
         echo date_to_string($date)."</div>";
         echo "<div id=\"selectGroup\">";
         echo "<form action=\"EditGroup.php\" method=\"post\">";
-        echo "<div id=\"list\"><ul>";
         $timecount = 0;
-        $sql = "SELECT * FROM Group_Schedule WHERE date = '$date'";
-        $record = $COMMON->executeQuery($sql, $_SERVER["SelectGroup.php"]);
         
-        for($i = 0; $i < mysql_num_rows($record); $i++){
-            $db_time = mysql_result($record, $i, 'time');
-            if(!is_group_null($date, $db_time)){
-                $time = short_time($db_time);
-                echo "<li>";
+        echo "<table id=\"selectGroupTable\">";
+       // echo "<tr><td><input type=\"radio\" name=\"time\" value=\"New\"><b>Add New</b></td></tr>"; 
+        foreach($apptTimes as $t){
+            $time = db_time($t);
+            $sql = "SELECT * FROM Group_Schedule WHERE date = '$date' and time='$time'";
+            $record = $COMMON->executeQuery($sql, $_SERVER["SelectGroup.php"]);
+            if(mysql_num_rows($record) == 1){
+                if(!is_group_null($date, $time)){
+                    $time = short_time($time);
+                    echo "<tr><td><b>";
+                    echo "<input type=\"radio\" name=\"time\" value=\"".$time."\">";
+                    echo display_time($time)."</b></td><td>Advisors: ";
+                    echo count_group_advisors($date, $time);
+                    echo "</td><td>Students: ".count_students($date, $time)."</td><td>";
+                    echo "Limit: ".get_size($date, $time)."</td></tr>";
+                    $timecount += 1;
+                }
+            }
+            //if time is not in db
+            else {
+                echo "<tr><td><b>";
                 echo "<input type=\"radio\" name=\"time\" value=\"".$time."\">";
-                echo display_time($time)."<div id=\"tab\">Advisors: ";
-                echo count_advisors($date, $time);
-                echo "<div id=\"tab\">Students: ".count_students($date, $time)."</li>";
-                $timecount += 1;
+                echo display_time($time)."</b></td><td>(Add time)</td>";
             }
         }
-        echo "</ul></div>";
+        echo "</table></div>";
         echo "<input type=\"hidden\" name=\"date\" value=\"".$date."\">";
-        echo "<div id=\"submit\"><input type=\"submit\" name=\"submitDay\"></div>";
+        echo "<div id=\"submit\"><input type=\"submit\" name=\"submitTime\"></div>";
     }
     echo "</form>";
 }
-
+else {
+    echo "<div id=\"error\">";
+    echo "<img src=\"includes/error.png\" id=\"errorImg\">";
+    echo "You are not logged in.</div>";
+}
 include_once 'includes/overallfooter.php';
 ?>
